requirements.txt
# =============================================
# DJANGO CORE
# =============================================
Django>=5.0,<6.0
django-environ>=0.11.0

# =============================================
# DATABASE
# =============================================
psycopg2-binary>=2.9.0

# =============================================
# API REST
# =============================================
djangorestframework>=3.15.0

# =============================================
# SECURITY & RATE LIMITING
# =============================================
django-ratelimit>=4.0.0
django-cors-headers>=4.3.0

# =============================================
# PRODUCTION SERVER
# =============================================
gunicorn>=21.2.0

# =============================================
# TESTING
# =============================================
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1

# =============================================
# CELERY (opzionale - decommenta se necessario)
# =============================================
# celery>=5.3.0
# redis>=5.0.0
# django-celery-beat>=2.5.0

# =============================================
# FILE STORAGE (opzionale - decommenta se necessario)
# =============================================
# Pillow>=10.1.0  # Per elaborazione immagini
# django-storages>=1.14.0  # Per S3/cloud storage
# boto3>=1.34.0  # AWS SDK per S3

# =============================================
# UTILITIES
# =============================================
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.3.0
